Move utils and components into src directory
