+++ src/components/SubNav/SubNav.tsx
@@ -1,5 +1,5 @@
-import RCLink from "../../components/RCLink/RCLink"
-import styles from "../../styles/components/SubNav.module.scss"
+import RCLink from "../RCLink/RCLink"
+import styles from "../../../styles/components/SubNav.module.scss"
 import { type RCPage } from "../../config/types"
 
 interface SubNavProps {
