+++ src/utils/bibUtils.ts
