+++ src/utils/searchUtils.ts
